services/fertilizer-application/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
pymongo==4.6.0
structlog==23.2.0
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Fertilizer application specific dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0

# Agricultural calculations
agricultural-calculations==0.1.0

# Database and caching
asyncpg==0.29.0
redis-py==5.0.1

# Logging and monitoring
loguru==0.7.2
sentry-sdk==1.38.0

# Development tools
black==23.9.1
flake8==6.1.0
mypy==1.6.1
pre-commit==3.5.0